+++ packages/parte-icons/src/Icons/ChevronRightIcon.tsx
@@ -1,0 +1,23 @@
+import { convertIcon, IconProps } from "../common/Icon";
+function Icon({ size, ...rest }: Pick<IconProps, "style" | "color" | "size">) {
+  return (
+    <svg
+      width={size}
+      height={size}
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      {...rest}
+    >
+      <path
+        d="M6.33301 4L10.333 8L6.33301 12"
+        stroke="currentcolor"
+        stroke-width="1.5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+    </svg>
+  );
+}
+const ChevronRightIcon = convertIcon(Icon, "chevron-right");
+export default ChevronRightIcon;